deployment/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
lime
shap
opencv-python